requirements.txt
mysql-connector-python
numpy
matplotlib==3.2.0
python_dotenv
tkcalendar